requirements.txt
flask==2.3.3
mysql-connector-python==8.0.33
gunicorn==21.2.0
boto3==1.28.62
requests==2.31.0
python-dotenv==1.0.0